refactor(ui): radio/checkbox size variant remove
