+++ src/playground/src/navigation/index.ts
@@ -1,5 +1,4 @@
 export {
   RouteType,
   Router,
-  routes,
 } from './navigation.component';
