+++ src/playground/src/navigation/navigation.component.tsx
@@ -32,30 +32,6 @@ export interface RouteType {
   name: string;
 }
 
-export const routes: RouteType[] = [
-  { name: 'Avatar' },
-  { name: 'Bottom Navigation' },
-  { name: 'Button' },
-  { name: 'Button Group' },
-  { name: 'Checkbox' },
-  { name: 'Dialog' },
-  { name: 'Input' },
-  { name: 'Layout' },
-  { name: 'List' },
-  { name: 'Popover' },
-  { name: 'Radio' },
-  { name: 'Radio Group' },
-  { name: 'Tab Bar' },
-  { name: 'Tab View' },
-  { name: 'Tooltip' },
-  { name: 'Text' },
-  { name: 'Toggle' },
-  { name: 'View Pager' },
-  { name: 'Top Navigation' },
-  { name: 'Modal' },
-  { name: 'Overflow Menu' },
-];
-
 const AppNavigator = createStackNavigator({
   ['Home']: Home,
   ['Avatar']: Avatar,
